% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_array.R
\name{create_lazyarray}
\alias{create_lazyarray}
\title{Create a lazy-array with given format and dimension}
\usage{
create_lazyarray(
  path,
  storage_format,
  dim,
  dimnames = NULL,
  compress_level = 50L,
  meta_name = "lazyarray.meta"
)
}
\arguments{
\item{path}{path to a local drive to store array data}

\item{storage_format}{data type, choices are \code{"double"}, 
\code{"integer"}, \code{"character"}, and \code{"complex"}}

\item{dim}{integer vector, dimension of array, see \code{\link{dim}}}

\item{dimnames}{list of vectors, names of each dimension, see \code{\link{dimnames}}}

\item{compress_level}{0 to 100, level of compression. 0 means
no compression, 100 means maximum compression. For persistent data,
it's recommended to set 100. Default is 50.}

\item{meta_name}{header file name, default is \code{"lazyarray.meta"}}
}
\value{
A \code{ClassLazyArray} instance
}
\description{
Create a directory to store lazy-array. The path must be missing. See \code{\link{load_lazyarray}} for more details
}
\details{
Lazy array stores array into hard drive, and load them on
demand. It uses multi-thread which gains significant 
speed boost on solid state drives. 

For examples, see \code{\link{lazyarray}}.
}
\author{
Zhengjia Wang
}
