% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs-lm.R
\name{lazy_lm}
\alias{lazy_lm}
\alias{lazy_lm_simple}
\title{Perform linear regression on \code{LazyMatrix} object}
\usage{
lazy_lm(
  formula,
  data,
  weights = NULL,
  na.action = "na.pass",
  center = FALSE,
  scale = FALSE,
  residuals = FALSE,
  ...
)

lazy_lm_simple(
  data,
  yidx,
  xidx,
  intercept = TRUE,
  weights = NULL,
  residuals = FALSE,
  ...
)
}
\arguments{
\item{formula}{R formula}

\item{data}{\code{LazyMatrix} object}

\item{weights}{weights of observations}

\item{na.action}{see \code{\link[stats]{lm}}}

\item{center, scale}{logical or numeric vectors with length equaling to 
data columns}

\item{residuals}{whether to return residuals}

\item{...}{passed to \code{\link{chunk_map}}}

\item{yidx}{column indices for response variable}

\item{xidx}{column indices for predictor}

\item{intercept}{whether to include intercept}
}
\value{
A \code{"lm"} or \code{"mlm"} object
}
\description{
Perform linear regression on \code{LazyMatrix} object
}
\examples{

a <- matrix(rnorm(100), ncol = 5)
x <- as.lazymatrix(a)


stats_lm <- stats::lm.fit(x = x[,-1], y = x[,1])$coefficients

lazy_lms <- lazy_lm_simple(x, yidx = 1, intercept = FALSE)
lazy_lms


# lazy_lm requires column names
colnames(x) <- c('y', 'x1', 'x2', 'x3', 'x4')
lazy_lm <- lazy_lm(y~.-1, data = x)
lazy_lm



}
