% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{lazy_parallel}
\alias{lazy_parallel}
\title{Schedule parallel processes for \code{LazyArray}}
\usage{
lazy_parallel(
  strategy = c("multisession", "multicore", "multiprocess", "cluster", "remote",
    "callr"),
  enabled = TRUE,
  workers = "auto",
  ...
)
}
\arguments{
\item{strategy}{strategies to apply, see \code{\link[future]{plan}} for
some of the details. For \code{"callr"} plan, please install package}

\item{enabled}{whether multiple-process strategy is enabled}

\item{workers}{positive integer or \code{"auto"}, number of 'CPU' to use.
The default value is \code{"auto"}, i.e. \code{future::availableCores()}}

\item{...}{Further passed to \code{\link[future]{plan}}}
}
\description{
Enable parallel processing, need \code{dipsaus} to be installed.
For \code{"callr"} strategy, please also install \code{future.callr}.
}
