% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{set_lazy_threads}
\alias{set_lazy_threads}
\title{Set Number of Threads for Lazy Arrays}
\usage{
set_lazy_threads(nr_of_threads = NULL, reset_after_fork = NULL)
}
\arguments{
\item{nr_of_threads}{number of CPU cores to use, or \code{NULL} to 
use all cores}

\item{reset_after_fork}{whether to reset after forked process}
}
\value{
Number of cores currently used.
}
\description{
A ported function from \code{\link[fstcore]{threads_fstlib}}.
}
\seealso{
\code{\link[fstcore]{threads_fstlib}}
}
