% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{typeof,AbstractLazyArray-method}
\alias{typeof,AbstractLazyArray-method}
\title{Type of \code{LazyArray}}
\usage{
\S4method{typeof}{AbstractLazyArray}(x)
}
\arguments{
\item{x}{a \code{LazyArray} or an R object}
}
\value{
The type of data stored in the input
}
\description{
Type of \code{LazyArray}
}
